Initial layout of page
Getting the boxes to still appear if there was nothing inside them was a pain because I wanted to make it so you could scroll your mouse wheel and the boxes would stay in place. I tried a lot of stuff, but eventually found the solution of setting a height and width on the center-container; this meant I could give the board-container a cell size of 33.3% instead of some hard coded px value (which would cause issues if you scrolled on the mouse wheel).

index.html
- Basic layout of page
 - Header will probably remain just the h1
 - Center-container has more work to do, the main part is the board-container which holds the actual board
 - post-board-container is where buttons will go later

styles.css
- Styling for the page
 - Generally, make most things a grid and sizing those grids
- The board itself
 - Sized by giving the center-container height: 90% and width: 80%, and then giving the board-container columns and rows of size 33.3%